splitted the codes into various files
