+++ src/player.js
@@ -1,0 +1,10 @@
+const Player = (name, sign) => {
+    const choices = [];
+    const addChoice = function (choice) {
+      choices.push(choice);
+    };
+    return {
+      name, sign, choices, addChoice,
+    };
+  }
+export default Player;
